feat: refactor android webview implementation
